add:处理星座数据列表

+++ src/main/java/com/tangerineSpecter/oms/system/service/system/MenuSettingService.java
@@ -17,6 +17,6 @@ public class MenuSettingService {
 
     public ServiceResult<Object> listInfo() {
         List<SystemMenu> list = systemMenuMapper.selectList(null);
-        return ServiceResult.pageSuccess(list, list.size());
+        return ServiceResult.pageSuccess(list, (long) list.size());
     }
 }
